+++ R/df2flextable.R
@@ -26,7 +26,30 @@ roundDf=function(df,digits=2){
      df
 }
 
-
+#'Change p value to string
+#'@param x A numeric vector
+#'@param digits integer indicating the number of decimal places
+#'@export
+#'@examples
+#'x=c(0.000001,NA,0.1234567,0.00123,0.000123)
+#'p2character(x)
+#'p2character(x,digits=4)
+p2character=function(x,digits=3){
+    # digits=3
+    cut=1/(10^digits)
+    temp=sprintf(paste0("%.",digits,"f"),x)
+    temp=stringr::str_replace(temp,"^0","")
+    temp
+    for(i in 1:length(x)){
+        if(is.na(x[i])){
+            temp[i]=""
+        } else if(x[i]<cut) {
+            temp[i]=stringr::str_replace(temp[i],"0$","1")
+            temp[i]=paste0("< ",temp[i])
+        }
+    }
+    temp
+}
 
 #' Convert data.frame to flextable
 #'
@@ -42,6 +65,7 @@ roundDf=function(df,digits=2){
 #' @param vlines Logical. Whether or not draw vertical lines
 #' @param colorheader Logical. Whether or not use color in header
 #' @param digits integer indicating the number of decimal places
+#' @param digitp integer indicating the number of decimal places of p values
 #' @param align_header alignment of header. Expected value is one of 'left', 'right', 'center', 'justify'.
 #' @param align_body alignment of body. Expected value is one of 'left', 'right', 'center', 'justify'.
 #' @param NA2space A logical. If true, convert NA value to space
@@ -59,13 +83,13 @@ roundDf=function(df,digits=2){
 #' df2flextable(head(iris),vanilla=TRUE,digits=c(1,2,3,4))
 #' df2flextable(head(iris),vanilla=FALSE)
 #' df2flextable(head(iris),vanilla=FALSE,vlines=FALSE,fontsize=14)
-#' df2flextable(head(mtcars))
+#' df2flextable(head(mtcars/2000),digits=3,pcol=8,digitp=4)
 #' }
 df2flextable=function(df,vanilla=FALSE,fontname=NULL,fontsize=12,
                       add.rownames=FALSE,
                       even_header="transparent",odd_header="#5B7778",
                       even_body="#EFEFEF",odd_body="transparent",
-                      vlines=TRUE,colorheader=FALSE,digits=2,
+                      vlines=TRUE,colorheader=FALSE,digits=2,digitp=3,
                       align_header="center",align_body="right",
                       NA2space=TRUE,pcol=NULL,...){
 
@@ -79,7 +103,15 @@ df2flextable=function(df,vanilla=FALSE,fontname=NULL,fontsize=12,
 
     if(!is.null(pcol)){
         for(i in 1:length(pcol)){
-            df[[pcol]][df[[pcol]]<0.001]="< 0.001"
+            df[[pcol[i]]]=p2character(df[[pcol[i]]],digitp)
+            # df[[pcol]]=sprintf(paste0("%.",digitp,"f"),df[[pcol]])
+            #
+            # temp=rep(0,digitp)
+            # temp=stringr::str_c(temp,collapse="")
+            # lookfor=paste0("0.",temp)
+            # rep=paste0("< 0.",stringr::str_c(rep(0,digitp-1),collapse=""),"1")
+            # rep
+            # df[[pcol]][df[[pcol]]==lookfor]=rep
         }
     }
 
@@ -192,7 +224,7 @@ df2flextable2=function(df,mincol=0.7,maxcol=4,...){
     # str(cwidth)
     # str(clen)
     ft=df2flextable(df,...) %>% width(width=cwidth) %>% align()
-    ft=df2flextable(df) %>% width(width=cwidth) %>% align()
+    # ft=df2flextable(df) %>% width(width=cwidth) %>% align()
     ft
     for(i in 1:ncol(df)){
         if(is.numeric(df[[i]])) ft<-ft %>% align(j=i,align="right")